hf-space/requirements.txt
gradio
fastapi
uvicorn
python-multipart
torch==2.1.2+cpu
torchvision==0.16.2+cpu
Pillow
gdown
numpy
